api/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
django-polymorphic==3.1.0
django-rest-polymorphic==0.1.10
django-cors-headers==4.3.1

psycopg2-binary==2.9.9
gunicorn==20.1.0